feat: answer generation step to async
The actual implementation doesn't give meaning full result on my
benchmark.
The other draw back is about the patch it self, I was constrain to
change the code here and there. We can do better for maintainability.
Last think, I think we should use the 'Routing Strategy' API from
[little](https://docs.litellm.ai/docs/routing#basic-reliability)
because it provide a way to manage the request per minute.

In overall, think again on the implementation to ease the batch
processing implementation.