refactor: code organisation + uniformization
In order to ease the implementation of another LLM API,
the parameters llm_names is remove in profit of 'LLMs_from_names' who built the
object needed. The future user will be able to implement they custom LLM classe and add to the list.

I choose to change a bit the file organisation in order to
ease the readability and manage future change easly.
The change:
- The abstract layer will be find in the folder 'src/ragtime/base'
- The specialization of thoses abstract class will be find under 'src/ragtime/generator/*'

This commit have also a small part about a feature - the config file -

WARNING: none of this commit is tested yet!!!

TODO:
- testing to check in there is not bug induce by the change
- documenting the breaking change